+++ Conditionals/Questions/src/com/company/AreaofTriangle.java
@@ -7,11 +7,11 @@ public class AreaofTriangle {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Enter the base value of triangle:- ");
-        int base = sc.nextInt();
+        float base = sc.nextFloat();
         System.out.println("Enter the height value of triangle:- ");
-        int height = sc.nextInt();
+        float height = sc.nextFloat();
 
-        int area = base * height / 2;
+        float area = base * height / 2;
         System.out.println("The area of an triangle is:- " + area);
     }
 }
